refactor: moves initial gameState to server
