+++ utils/context/GameContext.tsx
@@ -3,11 +3,11 @@
 import { createContext, ReactNode, use } from "react";
 import { createGameStore } from "../store/GameStore/gameStore";
 import { createGameStatisticsStore } from "../store/GameStatisticsStore/GameStatisticsStore";
+import { GameState } from "../store/GameStore/gameStore.types";
 
 interface Props {
   children: ReactNode;
-  synonymsOfDay: string[];
-  wordOfDay: string;
+  initialState: GameState;
   offsetDate: number;
 }
 interface GameContext {
@@ -17,16 +17,9 @@ interface GameContext {
 
 const GameContext = createContext<GameContext | null>(null);
 
-const GameProvider = ({
-  children,
-  offsetDate,
-  synonymsOfDay,
-  wordOfDay,
-}: Props) => {
+const GameProvider = ({ children, offsetDate, initialState }: Props) => {
   const gameStore = createGameStore({
-    synonyms: synonymsOfDay,
-    wordOfDay,
-    offsetDate,
+    initialState,
   });
   const gameStatisticsStore = createGameStatisticsStore(false, offsetDate);
